Refactors Instagram WebView for dark mode support
Moves Instagram web view logic to accommodate handling of dark mode settings.
This change includes removing the InstagramNavigationDelegate and inlining its logic, adds cookie support for setting dark mode preference.
It also adds a new initializer to accept ColorScheme? and determine dark mode.
